fix: renamed the css file
