+++ crowdin/messages.json
@@ -3181,6 +3181,15 @@
   "-115645240": "Please enter your mobile number.",
   "-381962759": "Please enter a valid email address.",
   "-1232661871": "Verification code is requried",
+  "-71049153": "Keep your account secure with a password",
+  "-981015173": "Go to Deriv Academy",
+  "-1193651304": "Country of residence",
+  "-2040322967": "Citizenship",
+  "-1541554430": "Next",
+  "-1770179687": "By clicking “Create account”, you confirm that you are 18 or older. You understand that we may use your email address to send you information about Deriv products and services as well as market news. You can always unsubscribe from these emails in your account settings. For more information, please take a look at Deriv’s <0>Security and privacy policy</0>.",
+  "-417711545": "Create account",
+  "-1264882776": "Your go-to platform for exclusive trading courses",
+  "-609665275": "Gain access to comprehensive modules on forex, Deriv MT5, and more.",
   "-142727028": "The email is in your spam folder (sometimes things get lost there).",
   "-276906320": "If you don't see an email from us within a few minutes, a few things could have happened:",
   "-1878043830": "Re-enter your email and try again",
@@ -3414,15 +3423,6 @@
   "-955173076": "We provide clear, accessible information about our platforms and processes so we can build trust and strong client relationships. Through detailed product descriptions and open communication, we aim to make it easier for you to make informed trading decisions.",
   "-2032112598": "Be responsible",
   "-35673135": "We build our services and platforms on a foundation of trust, honesty, and ethical practices. By proactively implementing policies and controls to prevent unlawful activities, we ensure a fair and secure environment for our trading community.",
-  "-71049153": "Keep your account secure with a password",
-  "-981015173": "Go to Deriv Academy",
-  "-1193651304": "Country of residence",
-  "-2040322967": "Citizenship",
-  "-1541554430": "Next",
-  "-1770179687": "By clicking “Create account”, you confirm that you are 18 or older. You understand that we may use your email address to send you information about Deriv products and services as well as market news. You can always unsubscribe from these emails in your account settings. For more information, please take a look at Deriv’s <0>Security and privacy policy</0>.",
-  "-417711545": "Create account",
-  "-1264882776": "Your go-to platform for exclusive trading courses",
-  "-609665275": "Gain access to comprehensive modules on forex, Deriv MT5, and more.",
   "-2125275828": "Unique trade types. Hundreds of instruments. Financial and derived markets.",
   "-154889447": "Ready to join over 2.5 million traders who have chosen Deriv as their trusted broker? Enter your email address to create a free demo account.",
   "-1957784093": "Easy And Free Sign Up | Online Trading | Deriv.com",
