+++ src/features/pages/academy-complete/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import { signup_wrapper, static_nav_logo } from '../signup-academy/signup-academy.module.scss'
 import Layout from 'features/components/templates/layout'
+import { nav_logo } from './signup-academy-complete.module.scss'
 import PopUpMenu from './pop-up-menu'
 import NavTemplate from 'features/components/templates/navigation/template'
 import LogoImage from 'images/common/deriv_academy.svg'
@@ -15,7 +15,7 @@ const SignupCompleteAcademy = () => {
                 has_centered_logo
                 renderLogo={() => (
                     <Link url={{ type: 'internal', to: '/' }}>
-                        <Image src={LogoImage} className={static_nav_logo} />
+                        <Image src={LogoImage} className={nav_logo} />
                     </Link>
                 )}
             />
