+++ src/features/pages/academy-complete/password-validation.tsx
