+++ src/features/pages/academy-signup/signup-academy.module.scss
@@ -1,40 +1,33 @@
 @use 'features/styles/theme/theme-mixins' as *;
 @import 'features/styles/theme/variables';
 
-.signup_wrapper {
+.wrapper {
     min-block-size: calc(100svh - 7rem);
 }
 
-.signup_form_container {
+.form_container {
     @include breakpoints(tablet) {
         box-shadow: 0px 0px 24px rgba(0, 0, 0, 0.05);
     }
 }
 
-.signup_form_line {
+.form_line {
     block-size: 1px;
     background-color: $color-grey-7;
 }
 
-.sign_up_header {
+.header {
     font-size: 24px;
 }
 
-.sign_up_content {
+.content {
     font-size: 12px;
 }
 
-.signup_login_redirect {
+.login_redirect {
     cursor: pointer;
 }
 
-.signup_fb_btn {
-    background-color: #1877f2;
-}
-
-.signup_apple_btn {
-    background-color: #000000;
-}
 .link_text {
     text-decoration: underline;
     &:hover {
@@ -42,11 +35,11 @@
         color: $text-color-primary;
     }
 }
-.static_nav_logo {
+.nav_logo {
     inline-size: 19.9rem;
     block-size: 3.19rem;
 }
-.signup_content {
+.content_new {
     max-width: 44rem;
 }
 .btn_signup {
