+++ src/features/pages/academy-signup/signup.content.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { signup_content } from './signup-academy.module.scss'
+import { content_new } from './signup-academy.module.scss'
 import Flex from 'features/components/atoms/flex-box'
 import Typography from 'features/components/atoms/typography'
 import { Localize } from 'components/localization'
@@ -13,10 +13,10 @@ const SignUpContent = () => {
             align="center"
             gap="8x"
         >
-            <Typography.Heading size="small" align="center" className={signup_content}>
+            <Typography.Heading size="small" align="center" className={content_new}>
                 <Localize translate_text="_t_Your go-to platform for exclusive trading courses_t_" />
             </Typography.Heading>
-            <Typography.Paragraph align="center" className={signup_content}>
+            <Typography.Paragraph align="center" className={content_new}>
                 <Localize translate_text="_t_Gain access to comprehensive modules on forex, Deriv MT5, and more._t_" />
             </Typography.Paragraph>
         </Flex.Box>
